Extract LiteralInterpreter.toExpression as LiteralEncoder class
